+++ components/eamxx/src/CMakeLists.txt
@@ -4,6 +4,7 @@ add_subdirectory(dynamics)
 add_subdirectory(physics)
 add_subdirectory(diagnostics)
 add_subdirectory(control)
+add_subdirectory(doubly-periodic)
 if (PROJECT_NAME STREQUAL "E3SM")
   add_subdirectory(mct_coupling)
 endif()
