+++ components/eamxx/src/physics/CMakeLists.txt
@@ -10,7 +10,6 @@ else()
   message(STATUS "WARNING: RRTMGP and COSP only supported for double precision builds; skipping")
 endif()
 add_subdirectory(shoc)
-add_subdirectory(dp)
 if (SCREAM_TEST_LEVEL GREATER_EQUAL SCREAM_TEST_LEVEL_EXPERIMENTAL)
   add_subdirectory(zm)
 endif()
